requirements.txt
scikit-learn>=1.4.0
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.11.0
streamlit>=1.35.0
plotly>=5.20.0
joblib>=1.3.0
pyyaml>=6.0
# Optional — install for CatBoost model support:
# catboost>=1.2.0
